+++ src/core/datatypes/forte_date_and_time.h
@@ -101,7 +101,7 @@ class CIEC_DATE_AND_TIME : public CIEC_ANY_DATE {
        *   \return number of bytes used in the buffer without trailing 0x00
        *           -1 on error
        */
-    virtual int toString(char* paValue, unsigned int paBufferSize) const ;
+    virtual int toString(char* paValue, size_t paBufferSize) const;
 
       /*! \brief Converts data type value to string - use GMT time
        *   \param paValue Pointer to the provided buffer
