+++ src/core/datatypes/forte_real.h
@@ -98,7 +98,7 @@ class CIEC_REAL : public CIEC_ANY_REAL{
      *   \return number of bytes used in the buffer without trailing 0x00
      *           -1 on error
      */
-    virtual int toString(char* paValue, unsigned int paBufferSize) const;
+    virtual int toString(char* paValue, size_t paBufferSize) const;
 
     /*! \brief Casts TForteFloat values to CIEC data types
      *  \param paSrcValue Source value for the cast
